Moved BFS to math.graph package.
