+++ src/nl/rubensten/utensils/io/Files.kt
@@ -13,10 +13,10 @@ import java.security.NoSuchAlgorithmException
  * @param other
  *          The file to compare the checksum of.
  * @param algorithm
- *          The hashing algorithm to use, defaults to SHA-256.
+ *          The hashing graph to use, defaults to SHA-256.
  * @return `true` if the files have the same checksum, `false` if not.
  * @throws IOException When files couldn't be found.
- * @throws NoSuchAlgorithmException When the algorithm is not available.
+ * @throws NoSuchAlgorithmException When the graph is not available.
  */
 @Throws(IOException::class)
 fun File.checksumEquals(other: File, algorithm: String = "SHA-256"): Boolean {
@@ -29,10 +29,10 @@ fun File.checksumEquals(other: File, algorithm: String = "SHA-256"): Boolean {
  * Generates the checksum of the file.
  *
  * @param algorithm
- *          The hashing algorithm to use, defaults to SHA-256.
+ *          The hashing graph to use, defaults to SHA-256.
  * @return The checksum of the file.
  * @throws IOException When there is a problem with the file I/O.
- * @throws NoSuchAlgorithmException When the algorithm is not available.
+ * @throws NoSuchAlgorithmException When the graph is not available.
  */
 @Throws(IOException::class)
 fun File.checksum(algorithm: String = "SHA-256"): String {
